bonus Update - fix somethings + Customer

+++ app/Http/Controllers/frontend/CartController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\frontend;
 
 
 use App\Models\Coupon;
+use App\Models\Customer;
 use Illuminate\Http\Request;
 use App\Http\Services\CartService;
 use Illuminate\Support\Facades\Session;
@@ -43,7 +44,7 @@ class CartController extends Controller
     {
         $this->cartService->update($request);
 
-        return redirect('/carts');
+        return redirect()->back();
     }
 
     public function remove($id = 0)
@@ -63,24 +64,31 @@ class CartController extends Controller
     public function applyCoupon(Request $request)
     {
         if ($request->has('coupon') && $request->coupon != '') {
+            $KhachHang =Customer::where('email',$request->email)->first();
+            if ($KhachHang) {
+                if ($KhachHang->coupon === $request->coupon) {
+                    return redirect('/carts')->with('error', 'Bạn đã sử dụng mã giảm giá này rồi!');
+                }}
             $onecoupon = Coupon::where('coupon', $request->coupon)->first();
-            Session::put(['name'=>$request->name]);
-            Session::put(['phone'=>$request->phone]);
-            Session::put(['address'=>$request->address]);
-            Session::put(['email'=>$request->email]);
-            Session::put(['contents'=>$request->contents]);
+            if($onecoupon) {
+                Session::put(['name' => $request->name]);
+                Session::put(['phone' => $request->phone]);
+                Session::put(['address' => $request->address]);
+                Session::put(['email' => $request->email]);
+                Session::put(['contents' => $request->contents]);
 
-                if ( $onecoupon->quantity > 0) {
+                if ($onecoupon->quantity > 0) {
                     $newquantity = $onecoupon->quantity - 1;
                     Coupon::where('coupon', $onecoupon->coupon)->update(['quantity' => $newquantity]);
-                    Session::put('coupon',[
-                        'macoupon'=> $onecoupon->coupon,
+                    Session::put('coupon', [
+                        'macoupon' => $onecoupon->coupon,
                         'gia' => $onecoupon->giam
                     ]);
                     return redirect()->back();
                 }
-            $onecoupon->delete();
-            return redirect('/carts')->with('error','Mã giảm giá đã hết!');
+                $onecoupon->delete();
+                return redirect('/carts')->with('error','Mã giảm giá đã hết!');
+            }
         }
 
         return redirect('/carts')->with('error','Mã giảm giá không đúng!');
